consolidate table and table-sortable
