+++ packages/components/addon/components/hds/table/index.js
@@ -1,32 +1,50 @@
 import Component from '@glimmer/component';
+import { action } from '@ember/object';
+import { guidFor } from '@ember/object/internals';
+import { tracked } from '@glimmer/tracking';
 
-/**
- * @class HdsTable
- */
 export default class HdsTableIndexComponent extends Component {
-  /**
-   * Sets the data source that gets yielded by the `:body` named block.
-   * @argument model
-   * @type {array}
-   */
-  /**
-   * Specifies which field the `@model` is sorted by. This **does not perform a sort**.
-   * It is only used to set states on the contextual `SortBy` components.
-   * @argument sortProperty
-   * @type {string}
-   */
-  /**
-   * Specifies which direction the `@model` is sorted. Descending when `true`, ascending when
-   * `false`. This **does not perform a sort**. It is only used to set states on the
-   * contextual `SortBy` components.
-   * @argument sortDescending
-   * @type {boolean}
-   */
+  a11yMessageId = 'table-a11y-message-' + guidFor(this);
 
-  // Convenience getter since all query params are strings.
-  get sortDescending() {
-    const arg = this.args.sortDescending;
-    if (typeof arg === 'boolean') return arg;
-    return arg === 'true';
+  @tracked sortBy = this.args.sortBy;
+  @tracked sortOrder = this.args.sortOrder || 'asc';
+  @tracked sortedMessageText = '';
+
+  get isSortableTable() {
+    const { columns } = this.args;
+    console.log('isSortableTable', columns !== undefined);
+    return columns !== undefined;
+  }
+
+  get columns() {
+    const { columns, sortingKeys } = this.args;
+
+    if (this.isSortableTable) {
+      columns.forEach((column) => {
+        // if we have a `sortingKeys` parameter only some columns can be sorted, otherwise all the columns are sortable
+        if (sortingKeys && sortingKeys.includes) {
+          column.isSortableColumn = sortingKeys.includes(column.key);
+        } else {
+          column.isSortableColumn = true;
+        }
+      });
+    }
+    return columns;
+  }
+
+  get getSortCriteria() {
+    return `${this.sortBy}:${this.sortOrder}`;
+  }
+
+  @action
+  setSortBy(column) {
+    if (this.sortBy === column) {
+      //invert the sort order
+      this.sortOrder = this.sortOrder === 'asc' ? 'desc' : 'asc';
+    } else {
+      this.sortBy = column;
+      this.sortOrder = 'asc';
+    }
+    this.sortedMessageText = `Sorted by ${this.sortBy} ${this.sortOrder}ending`;
   }
 }
