--- packages/components/addon/components/hds/table/table-sortable.hbs
@@ -1,21 +1,0 @@
-<table class="hds-table-sortable {{if @isStriped 'hds-table--striped'}}" data-test-table ...attributes>
-  <thead class="hds-table__thead" data-test-table-head>
-    <Hds::Table::Tr>
-      {{#each @columns as |column|}}
-        <Hds::Table::ThSort
-          @text={{column.label}}
-          @sortKey={{column.key}}
-          @sortBy={{this.sortBy}}
-          @sortOrder={{this.sortOrder}}
-          @setSortBy={{this.setSortBy}}
-        />
-      {{/each}}
-    </Hds::Table::Tr>
-  </thead>
-  <tbody class="hds-table__tbody" data-test-table-body>
-    {{#each (sort-by this.getSortCriteria @model) as |data|}}
-      {{yield data to="body"}}
-    {{/each}}
-  </tbody>
-</table>
-<div id={{this.a11yMessageId}} class="sr-only" role="region" aria-live="polite">{{this.sortedMessageText}}</div>
