--- packages/components/app/components/hds/table/table-sortable.js
@@ -1,1 +1,0 @@
-export { default } from '@hashicorp/design-system-components/components/hds/table/table-sortable';
