+++ packages/components/tests/dummy/app/templates/components/table.hbs
@@ -147,7 +147,7 @@ export default class ComponentsTableRoute extends Route {
   {{! template-lint-enable no-unbalanced-curlies }}
   {{! prettier-ignore-end }}
   <p class="dummy-paragraph">2. Invoke the
-    <span class="dummy-code">Hds::Table::TableSortable</span>
+    <span class="dummy-code">Hds::Table</span>
     component in your template file.</p>
   {{! prettier-ignore-start }}
   {{! template-lint-disable no-unbalanced-curlies }}
@@ -155,7 +155,7 @@ export default class ComponentsTableRoute extends Route {
     @language="handlebars"
     @code="<!-- app/templates/components/table.hbs -->
 
-<Hds::Table::TableSortable
+<Hds::Table
   @model=\{{this.model.data}}
   @sortBy=\{{this.sortBy}}
   @sortOrder=\{{this.sortOrder}}
@@ -168,7 +168,7 @@ export default class ComponentsTableRoute extends Route {
       <td>\{{folkData.year}}</td>
     </Hds::Table::Tr>
   </:body>
-</Hds::Table::TableSortable>"
+</Hds::Table>"
   />
 
 </section>
@@ -390,13 +390,10 @@ export default class ComponentsTableRoute extends Route {
       </Hds::Table::Tr>
     </:body>
   </Hds::Table>
-  {{! <h4 class="dummy-h4">
-    Static table with custom column widths
-  </h4> }}
   <h4 class="dummy-h4">
     Sortable table (all columns sortable)
   </h4>
-  <Hds::Table::TableSortable
+  <Hds::Table
     @model={{this.model.data}}
     @sortBy={{this.sortBy}}
     @sortOrder={{this.sortOrder}}
@@ -409,5 +406,5 @@ export default class ComponentsTableRoute extends Route {
         <td>{{data.year}}</td>
       </Hds::Table::Tr>
     </:body>
-  </Hds::Table::TableSortable>
+  </Hds::Table>
 </section>
