--- packages/components/tests/integration/components/hds/table/table-sortable-test.js
@@ -1,22 +1,0 @@
-import { module, test } from 'qunit';
-import { setupRenderingTest } from 'ember-qunit';
-import { render } from '@ember/test-helpers';
-import { hbs } from 'ember-cli-htmlbars';
-
-module('Integration | Component | hds/table/table-sortable', function (hooks) {
-  setupRenderingTest(hooks);
-
-  test('it renders', async function (assert) {
-    // Set any properties with this.set('myProperty', 'value');
-    // Handle any actions with this.set('myAction', function(val) { ... });
-
-    await render(hbs`<Hds::Table::TableSortable />`);
-
-    assert.dom(this.element).hasText('');
-  });
-
-  test('it should render with a CSS class that matches the component name', async function (assert) {
-    await render(hbs`<Hds::Table::TableSortable id="test-table-sort" />`);
-    assert.dom('#test-table-sort').hasClass('hds-table-sortable');
-  });
-});
